requirements.txt
asgiref==3.6.0
black==22.6.0
certifi==2022.12.7
click==8.0.4
colorama==0.4.6
Django==4.1.7
flit_core==3.6.0
mypy-extensions==0.4.3
pathspec==0.10.3
pip==22.3.1
platformdirs==2.5.2
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
python-decouple==3.8
setuptools==65.6.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
wheel==0.38.4
wincertstore==0.2
